+++ app/Http/Controllers/HorarioController.php
@@ -10,6 +10,7 @@ class HorarioController extends Controller
 {
     public function index($comisionId)
     {
+        
         $horarios = Horario::where('comision_id', $comisionId)->get();
         return response()->json($horarios);
     }
@@ -17,7 +18,7 @@ class HorarioController extends Controller
     public function store(Request $request, $comisionId)
     {
         $comision = Comision::findOrFail($comisionId);
-
+        $this->authorize('update', $comision);
         $validated = $request->validate([
             'dia_semana'  => 'required|in:lunes,martes,miercoles,jueves,viernes,sabado',
             'hora_inicio' => 'required|regex:/^\d{2}:\d{2}$/',
@@ -45,6 +46,7 @@ class HorarioController extends Controller
     {
         $horario = Horario::findOrFail($id);
         $comisionId = $horario->comision_id;
+        $this->authorize('update', $comision);
         $horario->delete();
 
         return redirect()->route('comisiones.show', $comisionId)
